fix(columbia_merge): refactor common funtions to be used in columbia merger and importer, harvard importer
update tests
update typing
update docstrings
